Added the Behavior cloning rollout logs for all 5 configs.
- The logs contained the rollout results for all 5 behavior cloning configs:
visuomotor-ultrafast, visuomotor, visuomotor-fast, ik-rel-fast and ik-rel.
- We basically validated the trained models with above configs over 20 test trials.
- Out of all 5 configs, only IK-rel-fast model was able to complete the kitchen task: picking up
the tomato can from the fridge and placing inside the microwave cavity centre. That too it was
able to do successfully only once out of 20 trials.
- Note we have tested using the pytorch model which had best validation loss during training, but still the results
are not satisfactory, as it implies with knowing just joint positions we cant successfully perform composite tasks.
- We also need a visual feedback, but even with visuomotor policy, success trials was zero, this might be due to the
incorrect placement of camera on robot`s arm, also during training we are not considering camera`s  depth information.
Though the visuomotor trained models exhibited human like natural movement  while moving near to the tomato can, but
due to above mentioned issues they were not able to pick and place tomato can properly. This might have caused the rollouts
to fail for visuomotor policy based models.
- This also affirms the statement that Imitation learning can be used just as a grounding factor to train much more stable
and efficient reinforcement policy. As our imitation learning model was trained merely on 300 random synthetic trials,
it might also be a reason that paucity of training data for IL model is causing the failures to happen during rollouts